requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
rank_bm25
datasets
peft
accelerate
tqdm